automations/requirements.txt
anthropic>=0.34.0
supabase>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.11
pytrends>=4.9.2
weasyprint>=62.0
jinja2>=3.1.4
pillow>=10.4.0
schedule>=1.2.2
